Update logic to run as validator where the developer can apply fixes himself
