--- .github/workflows/identifier_validation.yml
@@ -1,45 +1,0 @@
-name: Identifier Validation
-
-on:
-  pull_request:
-    branches:
-      - 'main'
-
-permissions:
-  contents: read
-
-jobs:
-  validate-identifiers:
-    name: Check New Identifier Findings
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout pull request
-        uses: actions/checkout@v6
-        with:
-          fetch-depth: 0
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v7
-        with:
-          enable-cache: true
-
-      - name: Set up Python
-        uses: actions/setup-python@v6
-        with:
-          python-version: '3.10'
-
-      - name: Install dependencies
-        run: uv sync --locked
-
-      - name: Extract base database
-        env:
-          BASE_SHA: ${{ github.event.pull_request.base.sha }}
-        run: |
-          mkdir -p /tmp/financedatabase-base
-          git archive "$BASE_SHA" database | tar -x -C /tmp/financedatabase-base
-
-      - name: Warn about new identifier findings
-        run: >-
-          uv run python scripts/validate_identifiers.py
-          --baseline /tmp/financedatabase-base/database
-          --github-annotations database
