+++ financedatabase/validation/__init__.py
@@ -1,0 +1,1 @@
+"""Development-only tooling for validating the source database. Excluded from the built package."""
