+++ pyproject.toml
@@ -61,10 +61,14 @@ build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
 packages = ["financedatabase"]
+exclude = ["financedatabase/validation"]
 
 [tool.hatch.build.targets.sdist]
 include = [
     "financedatabase/**",
+]
+exclude = [
+    "financedatabase/validation",
 ]
 
 [tool.black]
